Generates a website with HTML summaries for predictive models

Generates a website with HTML summaries for predictive models

modelDown(
  ...,
  modules = c("auditor", "drifter", "model_performance", "variable_importance",
    "variable_response"),
  output_folder = "output",
  repository_name = "repository",
  should_open_website = interactive()
)

Arguments

...
one or more explainers created with DALEX::explain() function. Pair of explainer could be provided to check drift of models

modules
modules that should be included in the website

output_folder
folder where the website will be saved

repository_name
name of local archivist repository that will be created

should_open_website
should generated website be automatically opened in default browser

Details

Additional arguments that could by passed by name:

remote_repository_path Path to remote repository that stores folder with archivist repository. If not provided, links to local repository will be shown.

device Device to use. Tested for "png" and "svg", but values from ggplot2::ggsave function should be working fine. Defaults to "png".

vr.vars variables which will be examined in Variable Response module. Defaults to all variables. Example vr.vars = c("var1", "var2")

vr.type types of examinations which will be conducteed in Variable Response module. Defaults to "partial". Example vr.type = c("partial", "conditional", "accumulated")

Author

Przemysław Biecek, Magda Tatarynowicz, Kamil Romaszko, Mateusz Urbański

Examples

# \donttest{
require("ranger")

#> Loading required package: ranger

require("breakDown")

#> Loading required package: breakDown

require("DALEX")

#> Loading required package: DALEX

#> Welcome to DALEX (version: 2.0).
#> Find examples and detailed introduction at: https://pbiecek.github.io/ema/
#> Additional features will be available after installation of: ggpubr.
#> Use 'install_dependencies()' to get all suggested dependencies

# Generate simple modelDown page
HR_data_selected <- HR_data[1000:3000,]
HR_glm_model <- glm(left~., HR_data_selected, family = "binomial")
explainer_glm <- explain(HR_glm_model, data=HR_data_selected, y = HR_data_selected$left)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  2001  rows  10  cols 
#>   -> target variable   :  2001  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.00244337 , mean =  0.5002499 , max =  0.9999995  
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9949444 , mean =  1.465729e-09 , max =  0.9772488  
#>   A new explainer has been created!  

modelDown::modelDown(explainer_glm,
                     modules = c("model_performance", "variable_importance",
                                 "variable_response"),
                     output_folder = tempdir(),
                     repository_name = "HR",
                     device = "png",
                     vr.vars= c("average_montly_hours"),
                     vr.type = "partial")

#> [1] "Generating model_performance..."

#> Saving 6.67 x 6.67 in image

#> Saving 6.67 x 6.67 in image

#> [1] "Generating variable_importance..."

#> Saving 6.67 x 6.67 in image

#> [1] "Generating variable_response..."

#> Saving 6.67 x 6.67 in image

# More complex example with all modules
HR_ranger_model <- ranger(as.factor(left) ~ .,
                      data = HR_data, num.trees = 500, classification = TRUE, probability = TRUE)
explainer_ranger <- explain(HR_ranger_model,
                      data = HR_data, y = HR_data$left, function(model, data) {
 return(predict(model, data)$prediction[,2])
}, na.rm=TRUE)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  14999  rows  10  cols 
#>   -> target variable   :  14999  values 
#>   -> predict function  :  function(model, data) {     return(predict(model, data)$prediction[, 2]) } 
#>   -> predicted values  :  numerical, min =  0 , mean =  0.2392687 , max =  1  
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7866413 , mean =  -0.0011862 , max =  0.7679504  
#>   A new explainer has been created!  

# Two glm models used for drift detection
HR_data1 <- HR_data[1:4000,]
HR_data2 <- HR_data[4000:nrow(HR_data),]
HR_glm_model1 <- glm(left~., HR_data1, family = "binomial")
HR_glm_model2 <- glm(left~., HR_data2, family = "binomial")
explainer_glm1 <- explain(HR_glm_model1, data=HR_data1, y = HR_data1$left)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  4000  rows  10  cols 
#>   -> target variable   :  4000  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.001281044 , mean =  0.5 , max =  1  
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9970616 , mean =  1.253318e-09 , max =  0.9891096  
#>   A new explainer has been created!  

explainer_glm2 <- explain(HR_glm_model2, data=HR_data2, y = HR_data2$left)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  11000  rows  10  cols 
#>   -> target variable   :  11000  values 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.000450076 , mean =  0.1428182 , max =  0.8093986  
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.8093986 , mean =  -4.519877e-12 , max =  0.9905684  
#>   A new explainer has been created!  

modelDown::modelDown(list(explainer_glm1, explainer_glm2),
  modules = c("auditor", "drifter", "model_performance", "variable_importance",
              "variable_response"),
  output_folder = tempdir(),
  repository_name = "HR",
  remote_repository_path = "some_user/remote_repo_name",
  device = "png",
  vr.vars= c("average_montly_hours", "time_spend_company"),
  vr.type = "partial")

#> [1] "Generating auditor..."

#> Warning: Module 'auditor' generation failed. Skipping it.The detailed error is: Error in xy.coords(x, y, xlabel, ylabel, log): 'x' is a list, but does not have components 'x' and 'y'

#> [1] "Generating drifter..."
#> Preparation of a new explainer is initiated
#>   -> model label       :  model_old 
#>   -> data              :  11000  rows  10  cols 
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> predict function  :  predict_function 
#>   -> predicted values  :  numerical, min =  0.0009104234 , mean =  0.4043219 , max =  1  
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> model_info        :  Model info detected classification task but 'y' is a NULL .  (  WARNING  )
#>   -> model_info        :  By deafult classification tasks supports only numercical 'y' parameter. 
#>   -> model_info        :  Consider changing to numerical vector with 0 and 1 values.
#>   -> model_info        :  Otherwise I will not be able to calculate residuals or loss function.
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  
#> Preparation of a new explainer is initiated
#>   -> model label       :  model_new 
#>   -> data              :  11000  rows  10  cols 
#>   -> target variable   :  not specified! (  WARNING  )
#>   -> predict function  :  predict_function 
#>   -> predicted values  :  numerical, min =  0.0009104234 , mean =  0.4043219 , max =  1  
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> model_info        :  Model info detected classification task but 'y' is a NULL .  (  WARNING  )
#>   -> model_info        :  By deafult classification tasks supports only numercical 'y' parameter. 
#>   -> model_info        :  Consider changing to numerical vector with 0 and 1 values.
#>   -> model_info        :  Otherwise I will not be able to calculate residuals or loss function.
#>   -> residual function :  difference between y and yhat (  default  )
#>   A new explainer has been created!  
#> [1] "Generating model_performance..."

#> Saving 6.67 x 6.67 in image

#> Saving 6.67 x 6.67 in image

#> [1] "Generating variable_importance..."

#> Saving 6.67 x 6.67 in image

#> [1] "Generating variable_response..."

#> Saving 6.67 x 6.67 in image

#> Saving 6.67 x 6.67 in image

# }